ai-sales-agent/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic-settings
python-dotenv
streamlit
requests
pytest
langchain-openai
langchain-core
langchain
firecrawl
